requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.4
django-coreapi==2.3.0
django-crontab==0.7.1
django-import-export==2.5.0
django-mathfilters==1.0.0
djangorestframework==3.12.2
et-xmlfile==1.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.0
Pillow==9.4.0
pytz==2022.7.1
PyYAML==6.0
requests==2.25.1
sqlparse==0.4.3
tablib==3.3.0
uritemplate==4.1.1
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
